migrate smartjson
